Initial Upload
This is the initial upload of the Swing Application Platform library.

This library contains one module, `gsul.application`, which has the following packages:

- `com.gsul.application`
   - `com.gsul.application.resources` (empty at this time)
   - `com.gsul.application.spi` (only has the skeleton `ModuleLoader` class)
   - `com.gsul.application.support` (the `MenuProvider` and `ToolbarProvider` classes and the `MenuType` and `ToolbarType` enums)
- `com.gsul.utils` (`GSLogger` has errors right now that will need to be fixed ***after*** we get the `ApplicaitonContext` modified for modular projects)

I figure that we can use the `com.gsul.application.resources` package to provide some icons that an Application Platform project can use if they want, but was not sure if we want to do that, which is why I left it empty.

Also, I am considering creating a `com.gsul.application.desktop` class in which the main window and standard options dialog could be located. After all, we are talking about making it to where modules can contribute to the application by providing menu items, toolbars, windows, dialogs, options panels, etc., so I do not think that it would be a bad idea to have these two windows and, possibly, the about dialog as part of the Application Platform library.

-SC