Deleted a Lot of Files
Cleaning out the repository because it will be easier for us to recreate the framework than to attempt to modify the original JSR-296 framework.

I removed all of the files under the `com.gsul.application` package for the above reason. I also deleted the `com.gsul.utils.GSLogger` because I have been reading about the `java.lang.System.Logger` and I think we may just want to use that. It is an interface, but then we could put `java.util.logging.Logger` behind it, or any other logging library we decide upon.

I modified the `module-info.java` file in the `gsul.application.framework` module to remove the exporting of the `com.gsul.application` package, since it is empty right now. I also modified the export of the `com.gsul.utils` package to only allow it to be required by the `gsul.application.platform` module.

Also, I created the `gsul.application.platform` module and set its `module-info.java` to require `java.base`, `java.desktop`, and `gsul.application.framework`.

The repository should now be ready to begin work on the new and improved Framework and Platform Projects...

-SC