Framework Started
I have started working on the framework module, since it will need to be created prior to creating the platform module.

The key to the whole framework/platform system is that the platform will rely upon the framework to take care of the basics of a GUI application project, while the framework will provide the plumbing for both CLI and GUI application projects. Therefore, we need to get the framework together and working first, then we can create the platform for use for more advanced GUI application projects. Simple GUI application projects will be able to use the framework directly.

-SC