Updates per Request
I have removed the two `java.beans` imports from the `Action` class, as they were not used, and added the `<build>` block to the Platform module, both as requested by @jkovalsky.

I have left the `requires java.logging;` requirement in the `module-info.java` file of the Framework module because it will be required by the `Application` class once I start writing the code for it, which is why I had included it in this PR.

Other than these changes, nothing else has been modified, so I will merge PR #1 and close it.

-SC