Removed Files and Requires
I have removed the JSR-296 Framework classes that I had started to include. This is because I am starting to think that it will be easier (or at least less complicated) to create our own Framework from scratch, rather than trying to modify the original JSR-296 Framework. The only original JSR-296 Framework class that I have kept is the `AbstractBean` class, because it may be useful moving forward. We can always delete it later if we don't need it.

I have also removed the `requires java.logging` line from the `module-info.java` file, as I am most likely going to put our utilities and custom logger back into the Framework project.

With these changes, I believe that we will be able to start thinking clearly about what this project needs and where it is going for the future. Also, in the Framework module, I believe that we should create a package called `com.gsul.framework.api` that contains mostly (if not entirely) interfaces that can be used by all implementing projects, including the Platform project.

-SC