Framework Complete (?)
I believe that the application framework portion of the library is complete at this point. However, I have yet to try testing it out. So, I am starting work on the `PlatformApplication` subclass of `Application`. As I get into this, we will find out if the framework is complete or not.

Also, I updated `StringUtils.wrap` to perform word wrapping and to take into account hard-coded newline characters and hyphenated words. Created a JUnit test for that method and the `StringUtils.insertTabLeader` method. Tests were run with 100% success (after quite a bit of fixing logic flaws in `StringUtils.wrap`). At this point, those two methods are 100% complete and ready to use, returning exactly what is expected based upon the parameters passed to them.

-SC