Application Framework Complete (?)
All classes have been cleaned up, but there is one `TODO` comment remaining because we need to decide as a team how to replace the use of `java.awt.Event` with `java.awt.AWTEvent`, because I could not see how to do this in my quick search.

At this point, I believe the framework portion of the platform is ready to be tested by creating a `PlatformApplication` in the `com.gs.platform` package. I guess that we will now see whether or not the `ResourceMap` and associated `Resource*` classes are going to play nice with the modular architecture that we are hoping to build.

If issues arise, we may need to look at the use of `Module` vs. the use of `ClassLoader`, as it currently is on the `ResourceMap` constructor.

-SC